Should complete the merge. Have static builds in the VC++ project since we can't be assed to do export macros for either platform. Note in readme the need for `dxgi` when it comes to linking against D3D. Change many of the types used to be more concrete than `unsigned int`; compile for x64 bit as well to reveal truncation casting errors from `std::size_t` to the integral types we return Prepare to refactor entire `instruction_set_t` enum into a set of flags that can be used to better identify features. TODO: 32-bit flag enum, enum flag support in codebase? TODO: really need to refactor GPU implementation. Both D3D and OpenGL and OpenCL offer extremely incomplete pictures of what the graphics cards have on it -- we need to go a lot deeper

+++ src/cpu/quantities_cache/windows.cpp
@@ -40,7 +40,7 @@ iware::cpu::quantities_t iware::cpu::quantities() {
 			case RelationProcessorCore:
 				++ret.physical;
 				// A hyperthreaded core supplies more than one logical processor.
-				ret.logical += std::bitset<sizeof(ULONG_PTR) * 8>(info.ProcessorMask).count();
+				ret.logical += static_cast<std::uint32_t>(std::bitset<sizeof(ULONG_PTR) * 8>(info.ProcessorMask).count());
 				break;
 
 			case RelationProcessorPackage:
